Create README.md
This is my Individual Project, where I use MSP (Message Send Protocol) using Python Language.
I created a one-way application from client to server. In this case, it is a sender-receiver application. 

This is the source code that I used.

I use this as my reference https://codinginfinite.com/python-chat-application-tutorial-source-code/
I look into Wikipedia to find the meaning of MSP.